Made a program to analyze the data and spit out a Latex document with the rezults but it is garbage. Some unknown changes were made to extract_values.py aswell
Changed retrieve_values.py to be able to store data from multiple files
in the form of a list(it can stll be used as before)

Added a class to angstroms_and_degrees.py that has a differencial
functions (only one currently but can add more later. This file will be
renamed in the future)

Made a New_analyze.py to replace the old one. It has a class that does
the same thing as the old one but does not yet generate a Latex file.
Everything else about it is better though

All the changes and new things besides analyze.py are compleatly
untested